refactoring: keep in one source code file.
