+++ core/rpc/core/src/impl/mod.rs
@@ -1,6 +1,4 @@
 mod build_adjust_account_tx;
-mod build_dao_tx;
-mod build_issue_udt_tx;
 mod build_transfer_tx;
 mod operation;
 mod query;
